Extract python_request_helpers.rs from turn.rs (parent #680)
Move the Python-specific request inspection helpers out of `turn.rs`
into a focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

Three production helpers extracted as free functions taking `&Agent`
(matching `actor_loop_flow` / `reply_retry` / earlier vertical-slice
precedent):

- `active_python_request_requires_tests` (`WorkMode::Python` + request
  explicitly requires tests)
- `python_verifier_available_for_requested_tests` (AutoTestRunner can
  detect a runnable Python test plan)
- `python_test_artifact_exists` (top-level `test_*.py` / `*_test.py` /
  `tests.py` scan of `work_root`)

Behavior unchanged: same `WorkMode::Python` gate, same `AutoTestKind::Test`
predicate, same top-level test-file pattern set.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `actor_loop_flow.rs` (4 sites) and `reply_retry.rs` (2 sites) to
the free-fn form.

LOC delta:
- turn.rs: 687 → 656 (-31)
- new module: +56

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
